web/MemeGenerator/app/requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
Pillow
requests
